% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.glht.R
\name{as_flextable.glht}
\alias{as_flextable.glht}
\title{Convert \link[multcomp]{glht} Object to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{glht}(x, ...)
}
\arguments{
\item{x}{\link[multcomp]{glht} object}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=as_flextable.glht]{as_flextable.glht()}} returns a \link[flextable]{flextable}.
}
\description{
..
}
\details{
Function \code{\link[=as_flextable.glht]{as_flextable.glht()}} add additional \link[flextable]{hline}(s)
for multiple \verb{$focus} in the input \link[multcomp]{glht} object.
}
\keyword{internal}
