% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glht.R
\name{glht}
\alias{glht}
\title{\link[multcomp]{glht} Object}
\description{
\link[multcomp]{glht} Object
}
\examples{
library(ecip); list(
 '`glht` via `aov`' = aov(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett')),
 '`glht` via `lm`, single `$focus`' = lm(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey')),
 '`glht` via `lm`, multiple `$focus`' = lm(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett'))
) |> fastmd::render2html(file = 'glht')

}
