% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glht.R
\name{md_.glht}
\alias{md_.glht}
\title{R Markdown Lines for \link[multcomp]{glht} Object}
\usage{
\method{md_}{glht}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[multcomp]{glht} Object
}
\examples{
library(ecip)

list(
 '`aov`' = aov(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett')),
 '`lm`, single `$focus`' = lm(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey')),
 '`glht` via `lm`, multiple `$focus`' = lm(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett'))
) |> fastmd::render_(file = 'glht')
}
\keyword{internal}
