% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glht.R
\name{S3_glht}
\alias{S3_glht}
\alias{confint_.glht}
\alias{confint_.confint.glht}
\alias{.pval.glht}
\alias{.pval.summary.glht}
\alias{.pval.summary.gtest}
\alias{.pval.mtest}
\alias{.pval.gtest}
\alias{estnm.glht}
\alias{expcoef.glht}
\alias{endpoint.glht}
\alias{nobsText.glht}
\title{Additional S3 Methods for \link[multcomp]{glht} Object}
\usage{
\method{confint_}{glht}(x, level = 0.95, ...)

\method{confint_}{confint.glht}(x, ...)

\method{.pval}{glht}(x)

\method{.pval}{summary.glht}(x)

\method{.pval}{summary.gtest}(x)

\method{.pval}{mtest}(x)

\method{.pval}{gtest}(x)

\method{estnm}{glht}(x)

\method{expcoef}{glht}(x)

\method{endpoint}{glht}(x)

\method{nobsText}{glht}(x)
}
\arguments{
\item{x}{a \link[multcomp]{glht} object}

\item{level}{..}

\item{...}{..}
}
\description{
Additional S3 Methods for \link[multcomp]{glht} Object
}
\examples{
library(multcomp)
m0 = lm(breaks ~ tension + wool, data = warpbreaks)
cp = mcp(tension = 'Tukey', wool = 'Dunnett')
m1 = aov(breaks ~ tension + wool, data = warpbreaks) |> glht(linfct = cp)
m2 = m0 |> glht(linfct = cp)
summary(m0)
summary(m1)
summary(m1, test = adjusted('none'))
.pval.glht(m1)

}
