% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glht.R
\name{glht}
\alias{glht}
\title{\link[multcomp]{glht} Object}
\description{
\link[multcomp]{glht} Object
}
\examples{
library(ecip); list(
 '`aov`' = aov(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett'))
) |> fastmd::render2html()

m = lm(breaks ~ tension + wool, data = warpbreaks)
library(ecip); list(
 '`lm`, single `$focus`' = m |> 
   glht(linfct = mcp(tension = 'Tukey')),
 '`lm`, multiple `$focus`' = m |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett'))
) |> fastmd::render2html()

}
