% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glht.R
\name{S3_glht}
\alias{S3_glht}
\alias{confint_.glht}
\alias{confint_.confint.glht}
\alias{.pval.summary.glht}
\alias{.pval.summary.gtest}
\alias{.pval.mtest}
\alias{.pval.gtest}
\alias{Sprintf.glht}
\title{Additional S3 Methods for \link[multcomp]{glht} Object}
\usage{
confint_.glht(x, level = 0.95, ...)

confint_.confint.glht(x, ...)

.pval.summary.glht(x)

.pval.summary.gtest(x)

.pval.mtest(x)

.pval.gtest(x)

Sprintf.glht(x)
}
\arguments{
\item{x}{a \link[multcomp]{glht} object}

\item{level}{..}

\item{...}{..}
}
\description{
Additional S3 Methods for \link[multcomp]{glht} Object
}
